# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puryaideu Api",
        "description": "Implementation of Swagger with in Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user_login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone": "+977 9843670972"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/user_logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\LoginController::logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user_detail": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\LoginController::user",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/specific_user": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\LoginController::user_id",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User details displayed successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/user_update/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\LoginController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "profile_pic": {
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Deepika",
                                    "email": "deepik@gmail.com",
                                    "phone": "+977 9812323132",
                                    "profile_pic": "2312.jpg"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Updated",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/history": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\LoginController::history",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}